refactor(dashboard): move to package apollo
