+++ sentinel-dashboard/src/main/java/com/alibaba/csp/sentinel/dashboard/apollo/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * save everything about apollo.
+ */
+package com.alibaba.csp.sentinel.dashboard.apollo;
