+++ sentinel-dashboard/src/main/java/com/alibaba/csp/sentinel/dashboard/apollo/util/DataSourceConverterUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alibaba.csp.sentinel.dashboard.util;
+package com.alibaba.csp.sentinel.dashboard.apollo.util;
 
 import com.alibaba.cloud.sentinel.datasource.RuleType;
 import com.alibaba.csp.sentinel.datasource.Converter;
